+++ app/Http/Controllers/CourseController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Course;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Yajra\DataTables\Facades\DataTables;
 use App\Http\Requests\StoreCourseRequest;
 use App\Http\Requests\UpdateCourseRequest;
@@ -17,6 +18,7 @@ class CourseController extends Controller
      */
     public function index()
     {
+        Gate::authorize('view', auth()->user());
         return view('setup.course.index');
     }
 
@@ -48,6 +50,7 @@ class CourseController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create', auth()->user());
         return view('setup.course.create');
     }
 
@@ -59,6 +62,7 @@ class CourseController extends Controller
      */
     public function store(StoreCourseRequest $request)
     {
+        Gate::authorize('create', auth()->user());
        $course = new Course();
        $course->name = $request->name;
        $course->price = $request->price;
@@ -76,6 +80,7 @@ class CourseController extends Controller
      */
     public function show(Course $course)
     {
+        Gate::authorize('view', auth()->user());
         return view('setup.course.show', compact('course'));
     }
 
@@ -87,6 +92,7 @@ class CourseController extends Controller
      */
     public function edit(Course $course)
     {
+        Gate::authorize('update', auth()->user());
         return view('setup.course.edit', compact('course'));
     }
 
@@ -99,6 +105,7 @@ class CourseController extends Controller
      */
     public function update(UpdateCourseRequest $request, Course $course)
     {
+        Gate::authorize('update', auth()->user());
         $course->name = $request->name;
         $course->price = $request->price;
         $course->description = $request->description;
@@ -115,6 +122,7 @@ class CourseController extends Controller
      */
     public function destroy(Course $course)
     {
+        Gate::authorize('delete', auth()->user());
         return $course->delete();
     }
 }
